Move mock_config_fixture to test_sensor
Since it is only used there
